update: da_job: reformatting unit tests + added integration tests
