+++ .env.example
@@ -2,7 +2,7 @@ HOST=
 PORT=
 DATABASE_URL=
 MADARA_RPC_URL=
-DA_LAYER=
+DA_LAYER=ethereum
 SETTLEMENT_LAYER=
 
 # Ethereum
